+++ bevy_asset_loader/examples/standard_material.rs
@@ -23,7 +23,7 @@ fn main() {
 
 #[derive(AssetCollection)]
 struct MyAssets {
-    #[asset(color_material)]
+    #[asset(standard_material)]
     #[asset(path = "images/player.png")]
     player: Handle<StandardMaterial>,
 }