+++ CHANGELOG.md
@@ -4,7 +4,7 @@
   - New example `no_loading_state`
 - rename derive attribute `color_material` to `standard_material`
 
-## v0.7.0
+## v0.8.0
 - update to Bevy version 0.6
 
 ## v0.7.0